% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Node & weights for a graph}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{the node of an edge}
  \item{v2}{the node of the other connecting edgeuse}
  \item{w}{weight of the edge}
  ...
}
}
\usage{
wiki_graph
}
\description{
A dataset containing node value and their respective weight and how they every node is connected to each other.
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\keyword{datasets}
